requirements.txt
biopython==1.84
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
greenlet==3.1.1
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
mysql-connector-python==9.1.0
networkx==3.4.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.36
tenacity==9.0.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
